requirements.txt
# Core UI + APIs
streamlit==1.34.0
python-dotenv==1.0.1
requests==2.31.0

# PDF & YouTube
PyMuPDF==1.26.0
youtube-transcript-api==0.6.1
pytube==15.0.0

# AI & NLP
langchain==0.1.14
tiktoken==0.5.1       # ✅ known to work on 3.12
openai==1.12.0

# Vector DB
faiss-cpu==1.7.4.post1  # ✅ this has wheels; don’t use just 1.7.4
